refactor(cypress): move events + editor into appBuilder; update refs + generator
events.js → appBuilder/events.js and editor/ → appBuilder/editor/ (both are
app-builder concerns). All import refs updated (0 dangling; webpack-resolution
smoke test green). generate-helper-docs file list repointed to the new
appBuilder/events.js + appBuilder/editor/textInput.js and index regenerated
(--check fresh).
