+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/components/csa.skip.js
@@ -5,7 +5,7 @@ import {
   selectCSA,
   selectEvent,
   addSupportCSAData,
-} from "Support/utils/events";
+} from "Support/utils/appBuilder/events";
 import {
   openAccordion,
   addDefaultEventHandler,
