+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/components/multiselectHappyPath.skip.js
@@ -36,7 +36,7 @@ import {
   selectCSA,
   selectEvent,
   addSupportCSAData,
-} from "Support/utils/events";
+} from "Support/utils/appBuilder/events";
 
 describe("Multiselect widget", () => {
   beforeEach(() => {
