+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/newSuits/componentsBasics/globalActions.skip.js
@@ -3,7 +3,7 @@ import { commonSelectors, commonWidgetSelector } from "Selectors/common";
 import { multipageSelector } from "Selectors/multipage";
 import {
     randomString
-} from "Support/utils/editor/textInput";
+} from "Support/utils/appBuilder/editor/textInput";
 
 import {
     addInputOnQueryField,
