+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/newSuits/componentsBasics/multiselect.skip.js
@@ -3,8 +3,8 @@ import { commonSelectors, commonWidgetSelector } from "Selectors/common";
 import {
     addCSA,
     verifyCSA
-} from "Support/utils/editor/textInput";
-import { addMultiEventsWithAlert } from "Support/utils/events";
+} from "Support/utils/appBuilder/editor/textInput";
+import { addMultiEventsWithAlert } from "Support/utils/appBuilder/events";
 import { openAndVerifyNode, openNode, verifyfunctions, verifyNodes, verifyNodeData } from "Support/utils/appBuilder/inspector";
