+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/newSuits/queries/chainingOfQueries.cy.js
@@ -1,8 +1,8 @@
 import { fake } from "Fixtures/fake";
 import { commonSelectors, commonWidgetSelector } from "Selectors/common";
 import { openEditorSidebar } from "Support/utils/commonWidget";
-import { selectEvent } from "Support/utils/events";
-import { randomString } from "Support/utils/editor/textInput";
+import { selectEvent } from "Support/utils/appBuilder/events";
+import { randomString } from "Support/utils/appBuilder/editor/textInput";
 import { buttonText } from "Texts/button";
 
 import {
