+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/newSuits/table/tableInteractions.cy.js
@@ -18,7 +18,7 @@ import {
   verifyTableExposedVars,
 } from "Support/utils/appBuilder/components/table";
 import { openEditorSidebar, openAccordion } from "Support/utils/commonWidget";
-import { addMultiEventsWithAlert } from "Support/utils/events";
+import { addMultiEventsWithAlert } from "Support/utils/appBuilder/events";
 import { resizeQueryPanel } from "Support/utils/appBuilder/querymanager/queryPanel";
 
 // ---------------------------------------------------------------------------
