+++ cypress-tests/cypress/support/utils/appBuilder/editor/codehinter.js
