+++ cypress-tests/cypress/support/utils/appBuilder/editor/editorHeaderOperations.js
