+++ cypress-tests/cypress/support/utils/appBuilder/editor/inputFieldUtils.js
