+++ cypress-tests/cypress/support/utils/appBuilder/editor/passwordNumberInput.js
@@ -8,7 +8,7 @@ import {
   addSupportCSAData,
   selectCSA,
   selectEvent,
-} from "Support/utils/events";
+} from "Support/utils/appBuilder/events";
 
 export const verifyControlComponentAction = (widgetName, value) => {
   cy.forceClickOnCanvas();
