+++ cypress-tests/cypress/support/utils/appBuilder/editor/text.js
