+++ cypress-tests/cypress/support/utils/appBuilder/events.js
